% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{add.extant.occ}
\alias{add.extant.occ}
\title{Add extant occurrence samples}
\usage{
add.extant.occ(tree, fossils, rho = 1)
}
\arguments{
\item{tree}{Phylo object.}

\item{fossils}{Fossils object.}

\item{rho}{Extant species sampling.}
}
\value{
An object of class fossils.
sp = edge labels. h = ages.
}
\description{
Add extant occurrence samples
}
\examples{
# simulate tree
lambda = 0.1
mu = 0.05
tips = 8
t<-TreeSim::sim.bd.taxa(tips, 1, lambda, mu)
t<-t[[1]]
# simulate fossils
f<-sim.fossils.poisson(t, 0.5, root.edge = FALSE)
# add extant samples
f<-add.extant.occ(t, f, rho = 0.5)
plot(f, t)
}
