% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{add.tip.samples}
\alias{add.tip.samples}
\title{Add extant and extinct tip samples}
\usage{
add.tip.samples(fossils, tree, species = NULL, rho = 1)
}
\arguments{
\item{fossils}{Fossils object.}

\item{tree}{Phylo object.}

\item{species}{Taxonomy object.}

\item{rho}{Tip sampling probability.}
}
\value{
An object of class fossils.
}
\description{
The tree is required to identify which edges are terminal.
}
\examples{
# simulate tree
t = ape::rtree(6)

# simulate fossils
f = sim.fossils.poisson(2, t)

# add tip samples
f = add.tip.samples(f, t, rho = 0.5)
plot(f, t)

}
