% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.sampling.R
\name{asymmetric.ages}
\alias{asymmetric.ages}
\title{Create a data.frame with asymmetric species durations}
\usage{
asymmetric.ages(tree, root.edge = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{root.edge}{If TRUE include root edge.}
}
\value{
Dataframe with internal asymmetric branch ages (min and max). Note that if root edge = TRUE the oldest lineage incorporates the origin.
}
\description{
Create a data.frame with asymmetric species durations
}
\examples{
t<-ape::rtree(6)
asymmetric.ages(t)
}

