% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{asymmetric.fossil.mapping}
\alias{asymmetric.fossil.mapping}
\title{Map asymmetric fossil lineages}
\usage{
asymmetric.fossil.mapping(tree, fossils)
}
\arguments{
\item{tree}{Phylo object.}

\item{fossils}{Fossils object.}
}
\value{
An object of class fossils.
}
\description{
Map fossils onto a tree assuming asymmetric (budding) speciation.
}
\examples{
# simulate tree
t = ape::rtree(6)

# simulate fossils
f = sim.fossils.poisson(tree = t, 2)

# add extant samples
f = add.extant.occ(f, tree = t, rho = 0.5)

# asymmetric mapping
f = asymmetric.fossil.mapping(t, f)

}

