% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.sampling.R
\name{asymmetric.fossil.samp}
\alias{asymmetric.fossil.samp}
\title{Sample asymmetric fossil lineages}
\usage{
asymmetric.fossil.samp(tree, fossils)
}
\arguments{
\item{tree}{Phylo object.}

\item{fossil}{Fossils object.}
}
\value{
An object of class fossils.
}
\description{
Sample fossils assuming asymmetric (budding) speciation.
}
\examples{
# simulate tree
t<-ape::rtree(6)
# simulate fossils
f<-sim.fossils.poisson(t, 2)
# add extant samples
f<-add.extant.occ(t, f, rho = 0.5)
# asymmetric sampling
f<-asymmetric.fossil.samp(t, f)
}

