% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.old.R
\name{attachment.times}
\alias{attachment.times}
\title{Identify attachment ages and extinction times in an incompletely sampled tree}
\usage{
attachment.times(tree, fossils, asymmetric.mapping = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{fossils}{Fossils object.}

\item{asymmetric.mapping}{If TRUE fossil sampling is assymmetric.
If speciation is "symmetric" and asymmetric.mapping = TRUE, the function calls asymmetric.fossil.mapping.
If speciation is "asymmetric" and asymmetric.mapping = FALSE, the function returns an error.}
}
\value{
Dataframe containing the speciation & extinction times in an incompletely sampled tree.
}
\description{
The age at which a species attaches to a tree may not be equivalent to the time of origin of a species
if sampling is incomplete.
This function takes an object of class fossils and the corresponding phylo object and calculates
the speciation (= attachment) times taking into account incomplete sampling.
}
\examples{
t<-ape::rtree(6)
# simulate fossils
f<-sim.fossils.poisson(t, 2)
# add extant samples
f<-add.extant.occ(t, f, rho = 0.5)
# asymmetric mapping
f<-asymmetric.fossil.mapping(t, f)
# calculate attachment times
attachment.times(t, f)
}

