% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{basin.age}
\alias{basin.age}
\title{Select a sensible basin age based tree age}
\usage{
basin.age(tree, root.edge = T)
}
\arguments{
\item{tree}{Phylo object.}

\item{root.edge}{Boolean indicating tree inclues a root edge.}
}
\value{
basin age
}
\description{
Select a sensible basin age based tree age
}

