% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{beast.fbd.format}
\alias{beast.fbd.format}
\title{Transforms a tree and fossils into a tree in Newick format which can be read by BEAST2.}
\usage{
beast.fbd.format(
  tree,
  fossils,
  rho = 1,
  sampled_tips = NULL,
  tip_order = "oldest_first",
  convert.to.ranges = FALSE,
  ...
)
}
\arguments{
\item{tree}{Complete tree.}

\item{fossils}{fossils dataframe.}

\item{rho}{Sampling probability of extant tips. Default 1, will be disregarded if sampled_tips is not null.}

\item{sampled_tips}{List of tip labels corresponding to sampled extant tips.}

\item{tip_order}{Order of indexes to assign to the tips, either `oldest_first` (by default, first = oldest and last = youngest) or `youngest_first` (reversed).}

\item{convert.to.ranges}{Whether to prune the fossils to only keep ranges, i.e. first and last appearances for each species.}

\item{...}{Additional parameters will be passed to ape::write.tree}
}
\value{
Output of write.tree.
}
\description{
Optionally, will convert fossil specimens to stratigraphic ranges - i.e., for each species, remove all the intermediate fossils and leave only the youngest and oldest.
The oldest and youngest of each species will be labelled according to the `tip_order` argument.
}
\examples{
# simulate tree
t = ape::rtree(6)

# simulate fossils
f = sim.fossils.poisson(rate = 2, tree = t)

# output for BEAST
beast.fbd.format(t, f) # output on the console
beast.fbd.format(t, f, convert.to.ranges = TRUE) # output on the console
\dontrun{
beast.fbd.format(t, f, file="example.tre") # output in file
beast.fbd.format(t, f, file="example.tre", convert.to.ranges = TRUE) # output in file
}
}
