% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{count.fossils.binned}
\alias{count.fossils.binned}
\title{Count the total number of fossils per interval}
\usage{
count.fossils.binned(fossils, interval.ages)
}
\arguments{
\item{fossils}{Fossils object.}

\item{interval.ages}{Vector of stratigraphic interval ages, starting with the minimum age of the youngest interval and ending with the maximum age of the oldest interval.}
}
\value{
Vector of extinct samples corresponding to each interval. Note the last value corresponds to the number of samples > the maximum age of the oldest interval.
}
\description{
Count the total number of fossils per interval
}

