% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count.fossils}
\alias{count.fossils}
\title{Count the total number of fossils}
\usage{
count.fossils(fossils, tol = 1e-08)
}
\arguments{
\item{fossils}{Fossils object.}

\item{tol}{Tolerance for identifying extant tips.}
}
\value{
Number of extinct samples.
}
\description{
Count the total number of fossils
}
