% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.sampling.R
\name{cryptic.specaition}
\alias{cryptic.specaition}
\title{Generate datasets of cryptic species}
\usage{
cryptic.specaition(ages, kappa)
}
\arguments{
\item{ages}{Dataframe of branching times for mixed species (asymmetric, symmetric, anagenic).}

\item{kappa}{Probability that a speciation event generates a cryptic species.}
}
\value{
Dataframe containing species labels, (morpho)species speciation & extinction times, mode of speciation, cryptic indicator, and corresponding cryptic speciation lables.
"a" = anagenic speciation, s" = symmetric speciation, "b" = budding (asymmetric) speciation, "o" = origin.
}
\description{
Generate datasets of cryptic species
}

