% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{cryptic.speciation}
\alias{cryptic.speciation}
\title{Generate datasets of cryptic species}
\usage{
cryptic.speciation(ages, kappa)
}
\arguments{
\item{ages}{Dataframe of branching times for mixed species (asymmetric, symmetric, anagenic).}

\item{kappa}{Probability that a speciation event generates a cryptic species.}
}
\value{
Dataframe containing species labels, (morpho)species speciation & extinction times, mode of speciation, cryptic indicator, and corresponding cryptic speciation lables.
"a" = anagenic speciation, s" = symmetric speciation, "b" = budding (asymmetric) speciation, "o" = origin.
}
\description{
Generate datasets of cryptic species
}
\examples{
# simulate tree
t<-ape::rtree(6)
# assign symmetric and asymmetric species
sp1<-mixed.ages(t, 0.5)
# simulate anagenic species
sp2<-anagenic.species(sp1, 0.1, parent.labels = T)
assign cryptic speciation events
sp3<-cryptic.speciation(sp2, 0.5)
}

