% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbdrange.R
\name{fbdrange}
\alias{fbdrange}
\title{Creates an fbdrange object from provided tree and data}
\usage{
fbdrange(tree, data)
}
\arguments{
\item{tree}{\code{phylo} object to be included in the fbdrange object}

\item{data}{associated data to be included in the fbdrange object. Needs to contain \code{range} and \code{orientation} fields.}
}
\value{
the new fbdrange object
}
\description{
Creates an fbdrange object from provided tree and data
}
