% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{fetch.descendants}
\alias{fetch.descendants}
\title{Fetch descendant lineages in a symmetric tree}
\usage{
fetch.descendants(edge, tree, return.edge.labels = F)
}
\arguments{
\item{edge}{Edge label.}

\item{tree}{Phylo object.}

\item{return.edge.labels}{If TRUE return all descendant edge labels instead of tips.}
}
\value{
Vector of symmetric descendants
}
\description{
Fetch descendant lineages in a symmetric tree
}
\examples{
t<-ape::rtree(6)
fetch.descendants(7,t)
fetch.descendants(7,t,return.edge.labels=TRUE)
}
