% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{format.for.beast}
\alias{format.for.beast}
\title{Transforms a tree and fossils into a sampled tree in beast-usable format and writes it in Newick format.
Designed to work with FBD.}
\usage{
\method{format}{for.beast}(tree, fossils, rho = 1, sampled_tips = NULL, ...)
}
\arguments{
\item{tree}{Complete tree.}

\item{fossils}{fossils dataframe.}

\item{rho}{Sampling probability of extant tips. Default 1, will be disregarded if sampled_tips is not null.}

\item{sampled_tips}{List of tip labels corresponding to sampled extant tips.}

\item{...}{Additional parameters will be passed to ape::write.tree}
}
\value{
Output of write.tree.
}
\description{
Transforms a tree and fossils into a sampled tree in beast-usable format and writes it in Newick format.
Designed to work with FBD.
}
\examples{
# simulate tree
t<-ape::rtree(6)
# simulate fossils
f<-sim.fossils.poisson(t, 2)
# output for BEAST
format.for.beast(t, f) # output on the console
format.for.beast(t, f, file="example.tre") # output in file
}
