% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossils.R
\name{fossils}
\alias{fossils}
\alias{as.fossils}
\alias{is.fossils}
\title{Fossils object}
\usage{
fossils(data, speciation.mode = c("asymmetric", "symmetric", "mixed"))

as.fossils(data, ...)

is.fossils(fossils)
}
\arguments{
\item{data}{Dataframe or list of sampled fossils. See Details for the list of required fields.}

\item{speciation.mode}{String indicating the speciation concept fossils were generated under (e.g. asymmetric/budding, symmetric/bifurcating, mixed). Defaults to mixed.}
}
\description{
Create a fossil record object from a dataframe or a list. The input is taken to be a dataframe or list.
}
\details{
The fossil record object contains 6 fields containing for each fossil the following data:
\itemize{
 \item \code{sp} the label of the corresponding species. This label matches with tip labels in the corresponding phylogeny
 and with species labels in the taxonomy record.
 \item \code{node} the label of the sampling node in the phylogeny, i.e the node at the end of the edge on which the fossil was sampled
 \item \code{origin} the label of the node where the corresponding species originated in the phylogeny
 \item \code{mode} the speciation mode that gave rise to the corresponding species. Can be one of \code{b} (budding/asymmetric), \code{s} (symmetric/bifurcation) or \code{a} (anagenic).
 \item \code{hmin} the lowest bound of the time interval in which that fossil was sampled
 \item \code{hmax} the highest bound of the time interval in which that fossil was sampled. Is equal to \code{hmin} if sampling times are exact.
}
}
