% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{fossilsToPaleotreeRecord}
\alias{fossilsToPaleotreeRecord}
\title{Transforms a tree, fossils dataframe and taxonomy (optional) into a fossilRecordSimulation object from package paleotree.}
\usage{
fossilsToPaleotreeRecord(tree, fossils, taxonomy = NULL, merge.cryptic = F)
}
\arguments{
\item{tree}{phylo object containing the tree}

\item{fossils}{fossils object}

\item{taxonomy}{optional taxonomy object. If NULL, all speciation is assumed symmetric}

\item{merge.cryptic}{whether cryptic species should be kept separate or merged, default FALSE}
}
\value{
The converted paleotree record
}
\description{
Transforms a tree, fossils dataframe and taxonomy (optional) into a fossilRecordSimulation object from package paleotree.
}

