% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling.R
\name{get.tip.descs}
\alias{get.tip.descs}
\title{Obtain the tips that define each node in a tree}
\usage{
get.tip.descs(tree)
}
\arguments{
\item{tree}{an object of class "Phylo".}
}
\value{
A list of vectors, with one entry for each node consisting of the tip labels
  that define that node.
}
\description{
Obtain the tips that define each node in a tree
}
\examples{
t = TreeSim::sim.bd.taxa(10, 1, 0.1, 0.05)[[1]]
get.tip.descs(t)
}
