% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbdrange.R
\name{get_fbdrange_from_file}
\alias{get_fbdrange_from_file}
\title{Import fbdrange object from file}
\usage{
get_fbdrange_from_file(input_file)
}
\arguments{
\item{input_file}{path to a tree file in Nexus format containing range and orientation metadata for the trees (for instance, a file generated by the BEAST2 package TODO)}
}
\value{
an fbdrange object
}
\description{
Import fbdrange object from file
}
\examples{
tree_file = system.file("extdata", "fbdrange.trees", package = "FossilSim")
fbdr = get_fbdrange_from_file(tree_file)
}
