% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{mixed.speciation}
\alias{mixed.speciation}
\title{Map asymmetric & symmetric lineages}
\usage{
mixed.speciation(tree, beta)
}
\arguments{
\item{tree}{Phylo object.}

\item{beta}{Probability of symmetric speciation.}
}
\value{
Dataframe of asymmetric or symmetric edge labels, parent edge labels and the mode of speciation.
"s" = symmetric speciation, "b" = budding (asymmetric) speciation, "o" = origin.
}
\description{
Asymmetric (or budding) speciation occurs with probability \eqn{\beta}. Asymmetric speciation gives rise to one new (morpho)species, while symmetric speciation gives rise to two new species and results in the extinction of the ancestor.
Note that if \eqn{\beta = 0} all speciation events will be asymmetric and if \eqn{\beta = 1} all speciation events will be symmetric.
}
\examples{
t<-ape::rtree(6)
mixed.speciation(t, 0.5)
}
