% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{paleotreeRecordToFossils}
\alias{paleotreeRecordToFossils}
\title{Transforms a fossilRecordSimulation object from package paleotree to a tree and fossils dataframe.}
\usage{
paleotreeRecordToFossils(record)
}
\arguments{
\item{record}{fossilRecordSimulation object.}
}
\value{
A list containing the converted tree and fossils
}
\description{
Transforms a fossilRecordSimulation object from package paleotree to a tree and fossils dataframe.
}
\examples{
# simulate record
record <- paleotree::simFossilRecord(p=0.1, q=0.1,r=0.1, nruns=1,nTotalTaxa=c(30,40), nExtant=0, nSamp = c(5,25))
# transform format
l_tf = paleotreeRecordToFossils(record)
l_tf$tree
l_tf$fossils
}

