% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phylo.fbd.R
\name{plot.phylo.fbd}
\alias{plot.phylo.fbd}
\title{plot.phylo.fbd: Plot object of class phylo.fbd}
\usage{
\method{plot}{phylo.fbd}(x, range.plot = FALSE, complete = FALSE, ...)
}
\arguments{
\item{range.plot}{Plot stratigraphic ranges?}

\item{complete}{Is this a complete FBD tree?}

\item{tree}{FBD tree to plot.}
}
\description{
plot.phylo.fbd: Plot object of class phylo.fbd
}
