% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.asymmetric.tree.R
\name{plot.SAtree}
\alias{plot.SAtree}
\title{Make a stratigraphic range plot from a tree object of class SAtree}
\usage{
\method{plot}{SAtree}(x, complete = FALSE, ...)
}
\arguments{
\item{x}{SAtree to plot.}

\item{complete}{Plot unsampled species.}

\item{...}{Additional parameters to be passed to \code{plot.default}.}
}
\description{
Make a stratigraphic range plot from a tree object of class SAtree
}
\examples{
tree = sim.fbd.taxa(10,1,3,2,1,TRUE)[[1]]
plot(SAtree(tree), complete=TRUE)

}
