% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling.R
\name{prune.fossils.to.ranges}
\alias{prune.fossils.to.ranges}
\title{Removes all intermediate fossils from a fossils object and labels the first and last samples of each species.}
\usage{
prune.fossils.to.ranges(fossils)
}
\arguments{
\item{fossils}{Fossils object.}
}
\value{
Updated fossils object, with intermediate fossils removed and an added `range` column
specifying if the fossil is a `first` or `last` appearance.
}
\description{
Extant samples will only be considered if included in the fossils object (using function `sim.extant.samples`).
If only one fossil is present for a particular species it is labelled as first.
}
\examples{
# simulate tree
t = ape::rtree(6)

# simulate fossils
f = sim.fossils.poisson(rate = 2, tree = t)

# add extant tips to the fossils object
f = sim.extant.samples(f, tree = t)

# prune fossils to keep only first and last appearances
fp = prune.fossils.to.ranges(f)

}
\seealso{
[prune.SAtree.to.ranges()] for the same function applied to a combined tree
}
