% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAtree.R
\name{prune.SAtree.to.ranges}
\alias{prune.SAtree.to.ranges}
\title{Removes all intermediate fossils from a combined tree and labels the first and last fossils of each species.}
\usage{
prune.SAtree.to.ranges(tree, taxonomy = NULL, fossils = NULL)
}
\arguments{
\item{tree}{Combined SAtree with fossils.}

\item{taxonomy}{Optional `taxonomy` object to update with the removed fossils.}

\item{fossils}{Options `fossils` object to update with the removed fossils and new labels.}
}
\value{
List of tree with pruned fossils, and updated taxonomy and fossils if provided.
}
\description{
First and last are based on the order used in the `\link{SAtree.from.fossils}` function, i.e. youngest first or oldest first.
Can be used with sampled or complete trees. If only one fossil is present for a particular species it is labelled as first.
If provided, the `taxonomy` and `fossils` objects will also be updated to match the pruned tree.
}
\examples{
# simulate tree
t = ape::rtree(6)

# simulate fossils
f = sim.fossils.poisson(rate = 2, tree = t)

# transform format
t2 = SAtree.from.fossils(t,f)$tree

# prune the tree to keep only ranges
t4 = prune.SAtree.to.ranges(t2)$tree

# or transform to sampled tree first
t3 = sampled.tree.from.combined(t2)$tree
t4 = prune.SAtree.to.ranges(t3)$tree
plot(t4)
}
\seealso{
[prune.fossils.to.ranges()] for the same function applied to a fossils object
}
