% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAtree.R
\name{SAtree}
\alias{SAtree}
\title{Tree with sampled ancestors represented as zero-length edges}
\usage{
SAtree(tree, complete = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{complete}{Whether the tree is complete. Default TRUE. If the tree is not complete, then all fossil tips correspond to fossil samples, otherwise only sampled ancestors are considered samples.}
}
\description{
Converts a phylo object to SAtree, without modification of tip labels.
}
