% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.taxonomy.R
\name{sim.anagenic.species}
\alias{sim.anagenic.species}
\title{Simulate anagenic species on a taxonomy object}
\usage{
sim.anagenic.species(tree, species, lambda.a)
}
\arguments{
\item{tree}{Phylo object.}

\item{species}{Taxonomy object.}

\item{lambda.a}{Rate of anagenic speciation. Default = 0.}
}
\value{
Object of class taxonomy.
}
\description{
Simulate anagenic species on a taxonomy object
}
\examples{
t = ape::rtree(10)
sp = create.taxonomy(t, 1)
add.anagenic.species(t, sp, 0.1)

}
\seealso{
\code{\link{taxonomy}}
}
