% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fbd.R
\name{sim.cdfbd.taxa}
\alias{sim.cdfbd.taxa}
\title{sim.cdfbd.taxa: Simulating Fossilized Birth Death Trees With Discrete Lineage-Specific Rate Regimes}
\usage{
sim.cdfbd.taxa(n, numbsim, lambda, mu, psi, rate, pi)
}
\arguments{
\item{n}{Number of extant sampled tips.}

\item{numbsim}{Number of trees to simulate.}

\item{lambda}{Vector of lineage-specific speciation rates.}

\item{mu}{Vector of lineage-specific extinction rates.}

\item{psi}{Vector of lineage-specific fossil sampling rates.}

\item{rate}{Rate of switching between rate regimes.}

\item{pi}{Root state frequencies.}
}
\value{
List of numbsim simulated trees with n extant sampled tips.
}
\description{
sim.cdfbd.taxa: Simulating Fossilized Birth Death Trees With Discrete Lineage-Specific Rate Regimes
}
\examples{
n<-10
numbsim<-1
# TODO sim.cdfbd.taxa(n,numbsim,c(2,1),c(0,0.3),c(1,0.1),1)
}
\keyword{birth}
\keyword{death}
\keyword{fossilized}

