% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.exponential}
\alias{sim.fossils.exponential}
\title{Simulate fossils under an exponential sampling model}
\usage{
sim.fossils.exponential(tree, rate, root.edge = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{rate}{Exponential sampling rate.}

\item{root.edge}{If TRUE include the root edge (default = TRUE).}
}
\value{
An object of class fossils.
sp = node labels. h = ages.
The label is for the node just below the sampled fossil.
}
\description{
Simulate fossils under an exponential sampling model
}
\examples{
# simulate tree
t<-ape::rtree(4)
# simulate fossils
rate = 2
f<-sim.fossils.exponential(t, rate)
plot(f, t)
}
\keyword{preservation}
\keyword{uniform}
