% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.non.unif.depth}
\alias{sim.fossils.non.unif.depth}
\title{Simulate fossils under a non-uniform model of preservation (Holland, 1995)}
\usage{
sim.fossils.non.unif.depth(tree, basin.age, strata, profile, PA = 0.5,
  PD = 0.5, DT = 0.5, root.edge = TRUE, convert.rate = FALSE)
}
\arguments{
\item{tree}{Phylo object.}

\item{basin.age}{Maximum age of the oldest stratigraphic interval.}

\item{strata}{Number of stratigraphic intervals.}

\item{PA}{Peak adbundance parameter.}

\item{PD}{Preferred depth parameter.}

\item{DT}{Depth tolerance parameter.}

\item{root.edge}{If TRUE include the root edge (default = TRUE).}

\item{convert.rate}{If TRUE convert per interval sampling probability into a per interval Poisson rate (default = FALSE).}
}
\value{
An object of class fossils.
sp = edge labels. h = fossil or interval ages. If convert.rate = TRUE, h = specimen age, if convert.rate = FALSE, h = max horizon age.
}
\description{
Simulate fossils under a non-uniform model of preservation (Holland, 1995)
}
\examples{
# simulate tree
t<-ape::rtree(6)
# assign a max age based on tree height
max = basin.age(t)
# generate water depth profile
strata = 4
wd<-sim.water.depth(strata)
# simulate fossils
f<-sim.fossils.non.unif.depth(t, max, strata, wd, PA = 1, PD = 0.5, DT = 1)
}
\keyword{fossil}
\keyword{non-uniform}
\keyword{preseravtion}

