% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.non.unif}
\alias{sim.fossils.non.unif}
\title{Simulate fossils under a non-uniform model of preservation}
\usage{
sim.fossils.non.unif(tree, interval.ages, sampling, root.edge = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{interval.ages}{Vector of stratigraphic interval ages, starting with the minimum age of the youngest interval and ending with the maximum age of the oldest interval.}

\item{sampling}{Vector of Poisson sampling rates. The first number corresponds to the youngest interval. The length of the vector should 1 less than the length of interval.ages.}

\item{root.edge}{If TRUE include the root edge (default = TRUE).}
}
\value{
An object of class fossils.
sp = edge labels. h = fossil ages.
}
\description{
Simulate fossils under a non-uniform model of preservation
}
\examples{
# simulate tree
t<-ape::rtree(6)
# assign a max age based on tree height
max = basin.age(t)
# assign interval times & rates
times = seq(0, max, length.out = 4)
rates = c(5, 3, 1)
# simulate fossils
f<-sim.fossils.non.unif(t, times, rates)
plot(f, t)
}
\keyword{fossil}
\keyword{non-uniform}
\keyword{preseravtion}

