% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.varible.R
\name{sim.fossils.non.unif.species}
\alias{sim.fossils.non.unif.species}
\title{Simulate fossils under a non-uniform model of preservation}
\usage{
sim.fossils.non.unif.species(tree, sampling = 1, variance = 1,
  interval.ages = NULL, root.edge = TRUE, return.sp.rates = TRUE)
}
\description{
Simulate fossils under a non-uniform model of preservation
}

