% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.poisson}
\alias{sim.fossils.poisson}
\title{Simulate fossils under a Poisson sampling model}
\usage{
sim.fossils.poisson(tree, phi, root.edge = T)
}
\arguments{
\item{tree}{Phylo object.}

\item{phi}{Poisson sampling rate.}

\item{root.edge}{If TRUE include the root edge (default = FALSE).}
}
\value{
dataframe of sampled fossils.
sp = edge labels. h = ages.
}
\description{
Simulate fossils under a Poisson sampling model
}
\examples{
t<-ape::rtree(4)
sim.fossils.poisson(t,1)
}
\keyword{preseravtion}
\keyword{uniform}

