% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.som.R
\name{sim.fossils.poisson.rde}
\alias{sim.fossils.poisson.rde}
\title{Simulate fossils under a Poisson sampling model rates varying for each edge and dependent on its ancestor}
\usage{
sim.fossils.poisson.rde(tree, rd = rlnorm(1, 0.1, 0.1), root.edge = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{rd}{The desired rate distribution from which the rates for individual edges would be
sampled (default = rlnorm(1,0.1,0.1)).}

\item{root.edge}{If TRUE include the root edge (default = TRUE).}
}
\value{
A list with an object of class fossils and a vector of rates applied on each edge.
sp = node labels. h = ages. rates = rates, The total number of rates is the sum of number of edges in the tree and root
edge
}
\description{
Simulate fossils under a Poisson sampling model rates varying for each edge and dependent on its ancestor
}
\examples{
# simulate tree
t<-TreeSim::sim.bd.taxa(4,1,1,0.10)[[1]]
# simulate fossils
f<-sim.fossils.poisson.rde(t)
plot(f$fossils, t)
f$rates
}
