% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils}
\alias{sim.fossils}
\title{Simulate fossils under a uniform model of preservation}
\usage{
sim.fossils()
}
\arguments{
\item{probability}{of preseravtion/sampling}
}
\description{
Simulate fossils under a uniform model of preservation
}
\examples{
sim.fossils()

}
\keyword{preseravtion}
\keyword{uniform}

