% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.species.rates}
\alias{sim.species.rates}
\title{Simulate fossils recovery rates with variation across lineages}
\usage{
sim.species.rates(rate = 1, tree = NULL, taxonomy = NULL,
  root.edge = TRUE, model = "autocorrelated", v = 0.01,
  dist = function() {     runif(1, 0, 2) }, jump.pr = 0.01,
  return.rates = TRUE)
}
\arguments{
\item{rate}{Initial rate at the origin or root of the phylo or taxonomy object. Default = 1.}

\item{tree}{Phylo object.}

\item{taxonomy}{Taxonomy object.}

\item{root.edge}{If TRUE include the root edge. Default = TRUE.}

\item{model}{Model used to simulate rate variation across lineages. Options include "autocorrelated" (default), "independent" or "jump".}

\item{v}{Brownian motion parameter \eqn{v} used in the autocorrelated rates model. Default = 0.01.}

\item{dist}{Distribution of rates used to draw new rates under the "independent" and "jump" models. This parameter is ignored if \code{model = "autocorrealted"}. The default is a uniform distribution with \emph{U(0, 2)}. The distribution function must return a single positive value.}

\item{jump.pr}{Probability that fossil recovery rate changes at speciation events. Default = 0.01.}
}
\value{
A vector of rates.
Rates are output for each species in the order they appear in the corresponding taxonomy object.
}
\description{
Available models include autocorrelated rates, independent rates and a "jump" model in which rate changes are coincident with speciation events.
Under the \code{autocorrelated} rates model, rates evolve along lineages according to a Brownian motion process, where the strength of the relationship between ancestor and descendant rates is determined by the parameter \eqn{\nu} (\code{v}).
If \eqn{\nu} is small rates will be more similar between ancestor and descendants, and if \eqn{\nu} is zero all rates will be equal.
For a given species \eqn{i} with ancestor \eqn{j}, a new rate \eqn{\kappa_i} is drawn from a lognornal distribution with
\deqn{\kappa_i = LN( \kappa_j - (\sigma^2/2), \sigma)}
where \eqn{\sigma = \nu * t_i} and \eqn{t_i} is the lineage duration of the species.
This fossil recovery model has been described in Heath et al. (2014) and is equivalent to the autocorrelated relaxed clock model described in Kishino et al. (2001).
Under the \code{independent} rates model a new rate is drawn for each species from any valid user-specified distribution (\code{dist}).
Under the \code{jump} model rates change at each speciation event with a given probability (\code{jump.pr}) and new rates are drawn from any valid user-specified distribution (\code{dist}).
}
\examples{
# simulate tree
t = ape::rtree(6)

# simulate taxonomy
s = sim.taxonomy(t, 0.5, 1, 0.5)

# simualte rates under the autocorrelated rates model
rate = 1
rates = sim.species.rates(rate = rate, taxonomy = s, v = 1)
f = sim.fossils.poisson(rates, species = s)
plot(f, t)

# simualte rates under the independent rates model
dist = function() { rlnorm(1, log(rate), 1) }
rates = sim.species.rates(rate = rate, taxonomy = s, model = "independent", dist = dist)
f = sim.fossils.poisson(rates, species = s)
plot(f, t)

# simualte rates under the jump model
rates = sim.species.rates(rate = rate, taxonomy = s, model = "jump", dist = dist, jump.pr = 0.1)
f = sim.fossils.poisson(rates, species = s)
plot(f, t)

}
\references{
Heath et al. 2014. The fossilized birth-death process for coherent calibration of divergence-time estimates. PNAS 111:E2957-E2966.\cr
Kishino et al. 2001. Performance of a divergence time Estimation method under a probabilistic model of rate evolution MBE 18:352–361.
}
