% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.simulation.R
\name{sim.uncertain.ages}
\alias{sim.uncertain.ages}
\title{Calculates minimum and maximum ages of fossils, based on stratigraphic intervals.
Allows a fraction of error on the correct stratigraphic interval.}
\usage{
sim.uncertain.ages(fossils, intervals, error_fraction = 0, error_range = 1,
  sd_intervals = NULL, save_params = F)
}
\arguments{
\item{fossils}{Fossils object.}

\item{intervals}{Stratigraphic intervals as a vector of interval boundaries in increasing order.}

\item{error_fraction}{Probability of error in interval identification per fossil. Default 0.}

\item{error_range}{Maximum deviation from the correct interval, expressed in number of intervals. Default 1.}

\item{sd_intervals}{Standard deviations on interval boundaries. If null, all deviations are assumed 0.}

\item{save_params}{Whether simulation parameters (error_fraction and error_range) should be saved as attributes in the returned object.}
}
\value{
Fossils object with min and max ages of fossils.
}
\description{
Calculates minimum and maximum ages of fossils, based on stratigraphic intervals.
Allows a fraction of error on the correct stratigraphic interval.
}
\examples{
# simulate tree
t = ape::rtree(6)
# simulate fossils
f = sim.fossils.poisson(rate = 2, tree = t)
# simulate uncertainty
f = sim.uncertain.ages(f, c(0,TreeSim::getx(t)))
}
