% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.water.depth}
\alias{sim.water.depth}
\title{Simulate water depth profile}
\usage{
sim.water.depth(strata, depth = 2, cycles = 2)
}
\arguments{
\item{strata}{Number of stratigraphic horizons.}
}
\value{
dataframe of sampled water depths.
}
\description{
Simulate water depth profile
}
\examples{
wd<-sim.water.depth(100)
plot(wd, type="l")
}
\keyword{fossil}
\keyword{non-uniform}
\keyword{preservation}

