% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{species.end}
\alias{species.end}
\title{Find a species' end (i.e extinction) time from a taxonomy object}
\usage{
species.end(species, taxonomy)
}
\arguments{
\item{species}{Species id (as written in \code{taxonomy$sp}).}

\item{taxonomy}{Taxonomy object.}
}
\value{
End time.
}
\description{
Find a species' end (i.e extinction) time from a taxonomy object
}
