% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{species.start}
\alias{species.start}
\title{Find a species' start (i.e speciation) time from a taxonomy object}
\usage{
species.start(species, taxonomy)
}
\arguments{
\item{species}{Species id (as written in \code{taxonomy$sp}).}

\item{taxonomy}{Taxonomy object.}
}
\value{
Start time.
}
\description{
Find a species' start (i.e speciation) time from a taxonomy object
}
