% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling.R
\name{subsample.fossils.youngest}
\alias{subsample.fossils.youngest}
\title{Obtain a subsample of fossil occurrences containing the youngest fossil
sample in each node of the tree.}
\usage{
subsample.fossils.youngest(fossils, tree, complete = TRUE)
}
\arguments{
\item{fossils}{an object of class "fossils" that corresponds to fossil
occurrences.}

\item{tree}{an object of class "Phylo", representing the tree upon which the
fossil occurrences were simulated.}

\item{complete}{logical, if TRUE the youngest sample from each clade in the
complete tree is returned, if FALSE the youngest sample from each clade in
the extant only counterpart tree is returned.}
}
\value{
an object of class "fossils" containing the subsampled fossil
  occurrences.
}
\description{
Obtain a subsample of fossil occurrences containing the youngest fossil
sample in each node of the tree.
}
