% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{print.taxonomy}
\alias{print.taxonomy}
\alias{summary.taxonomy}
\title{Display taxonomy object}
\usage{
\method{print}{taxonomy}(x, max.length = 50, round.x = 12, ...)

\method{summary}{taxonomy}(object, max.length = 50, round.x = 12,
  details = TRUE, ...)
}
\arguments{
\item{x}{Taxonomy object.}

\item{max.length}{Max number of rows to print out.}

\item{round.x}{Number of decimal places to be used for species and edge ages.}

\item{details}{If TRUE include summary statistics.}
}
\description{
Display taxonomy object
}
