% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.old.R
\name{symmetric.ages}
\alias{symmetric.ages}
\title{Create a data.frame with symmetric species durations (not node ages)}
\usage{
symmetric.ages(tree, root.edge = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{root.edge}{If TRUE include root edge. Root edge takes the root label.}
}
\value{
Dataframe with internal branch ages (min and max).
}
\description{
Create a data.frame with symmetric species durations (not node ages)
}
\examples{
t<-ape::rtree(6)
symmetric.ages(t)
}

