% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tree.max}
\alias{tree.max}
\title{Find the maximum age in a phylo object (root age or origin time)}
\usage{
tree.max(tree, root.edge = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{root.edge}{If TRUE include the root edge (default = TRUE).}
}
\value{
max age
}
\description{
Function returns the the root age or the origin time (if \code{root.edge = TRUE}).
}
\examples{
t = ape::rtree(6)
tree.max(t, root.edge = FALSE)

}
