% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{add.anagenic.species}
\alias{add.anagenic.species}
\title{Add anagenic species to a taxonomy object}
\usage{
add.anagenic.species(tree, species, lambda.a)
}
\arguments{
\item{tree}{Phylo object.}

\item{species}{Taxonomy object.}

\item{lambda.a}{Rate of anagenic speciation. Default = 0.}
}
\value{
Object of class taxonomy.
}
\description{
Add anagenic species to a taxonomy object
}
\examples{
t = ape::rtree(10)
sp = create.taxonomy(t, 1)
add.anagenic.species(t, sp, 0.1)

}
\seealso{
\code{\link{taxonomy}}
}
