% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{add.cryptic.species}
\alias{add.cryptic.species}
\title{Add cryptic species to a taxonomy object}
\usage{
add.cryptic.species(species, kappa)
}
\arguments{
\item{species}{Taxonomy object.}

\item{kappa}{Probability that speciation event is cryptic.}
}
\value{
An object of class taxonomy. Note the origin or root can not be cryptic.
}
\description{
Add cryptic species to a taxonomy object
}
\examples{
t = ape::rtree(10)
sp = create.taxonomy(t, 1)
add.cryptic.species(sp, 0.5)

}
\seealso{
\code{\link{taxonomy}}
}

