% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{add.extant.occ}
\alias{add.extant.occ}
\title{Add extant occurrence samples}
\usage{
add.extant.occ(fossils, tree = NULL, species = NULL, rho = 1,
  tol = NULL)
}
\arguments{
\item{fossils}{Fossils object.}

\item{tree}{Phylo object.}

\item{species}{Taxonomy object.}

\item{rho}{Extant species sampling probability.}

\item{tol}{Rounding error tolerance for tip ages.}
}
\value{
An object of class fossils.
}
\description{
Add extant occurrence samples
}
\examples{
# simulate tree
lambda = 0.1
mu = 0.05
tips = 8
t = TreeSim::sim.bd.taxa(tips, 1, lambda, mu)[[1]]

# simulate fossils
f = sim.fossils.poisson(0.5, t)

# add extant samples
f = add.extant.occ(f, t, rho = 0.5)
plot(f, t)

}
