% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{add.tip.samples}
\alias{add.tip.samples}
\title{Add extant and extinct tip samples}
\usage{
add.tip.samples(tree, fossils, rho = 1)
}
\arguments{
\item{tree}{Phylo object.}

\item{fossils}{Fossils object.}

\item{rho}{Tip species sampling.}
}
\value{
An object of class fossils.
sp = edge labels. h = ages.
}
\description{
Add extant and extinct tip samples
}
\examples{
# simulate tree
t<-ape::rtree(6)
f<-sim.fossils.poisson(t, 2)
# simulate fossils
# add tip samples
f<-add.tip.samples(t, f, rho = 0.5)
plot(f, t)
}
