% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.sampling.R
\name{anagenic.species}
\alias{anagenic.species}
\title{Simulate anagenic species along branches}
\usage{
anagenic.species(ages, lambda.a = 0.1, parent.labels = FALSE)
}
\arguments{
\item{ages}{Dataframe of branching times (for asymmetric and / or symmetric species).}

\item{lambda.a}{Rate of anagenic speciation.}

\item{parent.labels}{If TRUE print out parent edge labels.}
}
\value{
Dataframe containing species labels, (morpho)species speciation & extinction times, and mode of speciation.
"a" = anagenic speciation, s" = symmetric speciation, "b" = budding (asymmetric) speciation, "o" = origin.
}
\description{
Simulate anagenic species along branches
}
\examples{
t<-ape::rtree(6)
sp<-mixed.ages(t, 0.5)
sp<-anagenic.species(sp, 0.1)
}

