% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.times.R
\name{attachment.times}
\alias{attachment.times}
\title{Identify asymmetric attachment ages and extinction times in an incompletely sampled tree}
\usage{
attachment.times(tree, fossils)
}
\arguments{
\item{tree}{Phylo object.}

\item{fossils}{Fossils object.}
}
\value{
Dataframe containing the speciation & extinction times in an incompletely sampled tree.
}
\description{
The age at which a species attaches to a tree may not be equivalent to the time of origin of a species
if sampling is incomplete.
This function takes an object of class fossils and the corresponding phylo and calculates
the speciation (= attachment) times taking into account incomplete sampling.
The function assumes all speciation events are asymmetric (budding).
If the fossil object does not represent asymmetric species, the function converts species to asymmetric using the corresponding tree.
}
\examples{

# simulate tree
t<-ape::rtree(6)

# simulate fossils
f<-sim.fossils.poisson(2, t)

# add extant samples?
#f<-add.extant.occ(t, f, rho = 0.5)

# calculate attachment times
attachment.times(t, f)

}
