% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{basin.age}
\alias{basin.age}
\title{Define a basin age based on tree height}
\usage{
basin.age(tree, root.edge = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{root.edge}{If TRUE include the root edge (default = TRUE).}
}
\value{
basin age
}
\description{
Function returns an age slightly older than the root.age or origin time using the formula
\eqn{round(max,1) + 0.1}, where max is the root.age or origin time.
}
\examples{
t<-ape::rtree(6)
basin.age(t, root.edge = FALSE)
}

