% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{count.fossils}
\alias{count.fossils}
\title{Count the total number of fossils}
\usage{
count.fossils(fossils)
}
\arguments{
\item{fossils}{Fossils object.}
}
\value{
Number of extinct samples.
}
\description{
Count the total number of fossils
}

