% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fossils.R
\name{draw.fossils}
\alias{draw.fossils}
\title{Plot simulated fossils}
\usage{
draw.fossils(x, fossils, root.edge = FALSE, show.tip.label = TRUE,
  align.tip.label = FALSE, add.fossils = TRUE, ...)
}
\arguments{
\item{fossils}{Dataframe of sampled fossils (sp = edge labels. h = ages.)}

\item{root.edge}{Boolean indicating tree inclues a root edge.}

\item{add.fossils}{Boolean indicating plot should inclue fossils.}

\item{tree}{Phylo object.}
}
\description{
Plot simulated fossils
}
\examples{
t<-ape::rtree(10)
f<-sim.fossils.poisson(t,3)
}

