% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossils.R
\name{fossils}
\alias{fossils}
\alias{as.fossils}
\alias{is.fossils}
\title{Fossils object}
\usage{
fossils(data = NULL, from.taxonomy = FALSE)

as.fossils(data, ...)

is.fossils(fossils)
}
\arguments{
\item{data}{Dataframe or list of sampled fossils. See Details for the list of required fields. If NULL, creates a blank fossils object.}

\item{from.taxonomy}{Boolean indicating whether the fossils were sampled using a taxonomy object. Defaults to FALSE.}
}
\description{
Create a fossil record object from a dataframe or a list. The input is taken to be a dataframe or list.
}
\details{
The fossil record object contains 5 fields containing for each fossil the following data:
\itemize{
 \item \code{sp} the label of the corresponding species. This label matches with tip labels in the corresponding phylogeny
 and with species labels in the taxonomy record.
 \item \code{edge} the label of the sampling node in the phylogeny, i.e the node at the end of the edge on which the fossil was sampled
 \item \code{origin} the label of the node where the corresponding species originated in the phylogeny
 \item \code{hmin} the lowest bound of the time interval in which that fossil was sampled
 \item \code{hmax} the highest bound of the time interval in which that fossil was sampled. Is equal to \code{hmin} if sampling times are exact.
}
}
