% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossils.R
\name{fossils}
\alias{fossils}
\alias{as.fossils}
\alias{is.fossils}
\title{Fossils object}
\usage{
fossils(data, ages = NULL, speciation.mode = NULL)

as.fossils(data, ...)

is.fossils(fossils)
}
\arguments{
\item{data}{Dataframe or list of sampled fossils (sp = edge labels, h = ages).}

\item{ages}{String indicating how fossil ages are defined relative to interval ages (e.g. continuous, interval.max)}

\item{speciation.mode}{String indicating the speciation concept fossils were generated under (e.g. asymmetric/budding, symmetric/bifurcating, mixed)}
}
\description{
Create a fossil record object. The input is taken to be a dataframe or list with phylo object edge labels (sp) and sampled fossil ages (h).
}
