% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapping.R
\name{mixed.ages}
\alias{mixed.ages}
\title{Print out a data.frame with mixed speciation (asymmetric and symmetric) ages}
\usage{
mixed.ages(tree, beta, root.edge = T)
}
\arguments{
\item{tree}{Phylo object.}

\item{beta}{Probability of symmetric speciation.}

\item{root.edge}{If TRUE include root edge.}
}
\value{
Dataframe with internal asymmetric or symmetric branch ages (min and max). Note that if root edge = T the oldest lineage incorporates the origin.
"s" = symmetric speciation, "b" = budding (asymmetric) speciation, "o" = origin.
}
\description{
Function uses mixed.speciation() to assign asymmetric & symmetric lineages.
}
\examples{
t<-ape::rtree(6)
mixed.ages(t, 0.5)
}

