% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.sampling.R
\name{mixed.speciation}
\alias{mixed.speciation}
\title{Sample asymmetric & symmetric lineages}
\usage{
mixed.speciation(tree, f)
}
\arguments{
\item{tree}{Phylo object.}

\item{f}{Probability of asymmetric speciation.}
}
\value{
Dataframe of asymmetric or symmetric edge labels, parent edge labels and the mode of speciation.
"s" = symmetric speciation, "b" = budding (asymmetric) speciation, "o" = origin.
}
\description{
Asymmetric (or budding) speciation occurs with probability \eqn{f}. Asymmetric specaition gives rise to one new (morpho)species, while symmetric speciation gives rise to two new species and results in the extinction of the ancestor.
Note that if \eqn{f = 1} all speciation events will be asymmetric and if \eqn{f = 0} all speciation events will be symmetric.
}
\examples{
t<-ape::rtree(6)
mixed.speciation(t, 0.5)
}

