% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phylo.fbd.R
\name{plot.phylo.fbd}
\alias{plot.phylo.fbd}
\title{plot.phylo.fbd: Plot object of class phylo.fbd}
\usage{
\method{plot}{phylo.fbd}(tree, complete = FALSE)
}
\arguments{
\item{tree}{FBD tree to plot.}

\item{complete}{Is this a complete FBD tree?}
}
\description{
plot.phylo.fbd: Plot object of class phylo.fbd
}
\keyword{birth}
\keyword{death}
\keyword{fossilized}
