% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.plot.R
\name{rangeplot}
\alias{rangeplot}
\title{rangeplot: Make a stratigaphic range plot from an object of class phylo.fbd}
\usage{
rangeplot(tree, complete = FALSE)
}
\arguments{
\item{tree}{phylo.fbd object to plot.}

\item{complete}{Plot completely unsampled species?}
}
\description{
rangeplot: Make a stratigaphic range plot from an object of class phylo.fbd
}
