% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeplot.R
\name{rangeplot}
\alias{rangeplot}
\title{rangeplot: Make a stratigaphic range plot from an object of class phylo.fbd}
\usage{
rangeplot(x, complete = FALSE)
}
\arguments{
\item{x}{phylo.fbd object to plot.}

\item{complete}{Plot unsampled species?}
}
\description{
rangeplot: Make a stratigaphic range plot from an object of class phylo.fbd
}
\examples{
tree = sim.fbd.taxa(10,1,3,2,1,TRUE)[[1]]
rangeplot(tree, complete=TRUE)
}
