% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.assignment.R
\name{reassign.ages}
\alias{reassign.ages}
\title{Reassign fossil ages}
\usage{
reassign.ages(tree, fossils, basin.age, strata, poisson = FALSE,
  root.edge = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{fossils}{Fossils object.}

\item{basin.age}{Maximum age of the oldest horizon.}

\item{strata}{Number of stratigraphic intervals.}

\item{poisson}{If TRUE the function returns random ages.}
}
\value{
Fossils object with reassigned fossil ages.
sp = edge labels. h = ages.
}
\description{
Reassign fossil ages using the median age of a stratigraphic interval
or a random time point drawn from that interval (poisson = TRUE).
If a given species is only extant for a portion of the total interval duration
ages are reassigned using the median age of the branch duration
or a random time point drawn from that duration (poisson = TRUE).
}
\examples{
# simulate tree
t<-ape::rtree(8)
# assign a max age based on tree height
max = basin.age(t)
# simulate fossils
strata = 5
sampling = 1
f<-sim.fossils.unif(t,max,strata,sampling)
# reassign ages
reassign.ages(t,f,max,strata)
}

