% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fbd.R
\name{sim.fbd.age}
\alias{sim.fbd.age}
\title{sim.fbd.age: Simulating fossilized birth-death trees of a fixed age.}
\usage{
sim.fbd.age(age, numbsim, lambda, mu, psi, complete = FALSE)
}
\arguments{
\item{age}{Time since origin.}

\item{numbsim}{Number of trees to simulate.}

\item{lambda}{Speciation rate.}

\item{mu}{Extinction rate.}

\item{psi}{Fossil sampling rate.}

\item{complete}{Keep unsampled extinct lineages?}
}
\value{
Array of 'numbsim' trees with the time since origin being 'age'.
If tree goes extinct or no tips are sampled, return value is '0'.
If only one extant tip is sampled, return value is '1'.
}
\description{
sim.fbd.age: Simulating fossilized birth-death trees of a fixed age.
}
\examples{
age = 1
lambda = 2.0
mu = 0.5
psi = 0.6
numbsim = 2
sim.fbd.age(age, numbsim, lambda, mu, psi)
#TODO function sometimes returns at error
}
\keyword{birth}
\keyword{death}
\keyword{fossilized}
