% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fbd.R
\name{sim.fbd.taxa}
\alias{sim.fbd.taxa}
\title{sim.fbd.taxa: Simulating Fossilized Birth-Death Trees On A Fixed Number Of Extant Taxa.}
\usage{
sim.fbd.taxa(n, numbsim, lambda, mu, psi)
}
\arguments{
\item{n}{Number of extant sampled tips.}

\item{numbsim}{Number of trees to simulate.}

\item{lambda}{Speciation rate.}

\item{mu}{Extinction rate.}

\item{psi}{Fossil sampling rate.}
}
\value{
List of numbsim simulated trees with n extant sampled tips.
}
\description{
sim.fbd.taxa: Simulating Fossilized Birth-Death Trees On A Fixed Number Of Extant Taxa.
}
\examples{
n<-10
lambda <- 2.0
mu <- 0.5
psi <-0.6
numbsim<-2
# TODO sim.fbd.taxa(n,numbsim,lambda,mu,psi)
}
\keyword{birth}
\keyword{death}
\keyword{fossilized}
