% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.intervals}
\alias{sim.fossils.intervals}
\title{Simulate fossils under a non-uniform model of preservation for a given set of consecutive time intervals}
\usage{
sim.fossils.intervals(tree = NULL, species = NULL, interval.ages = NULL,
  basin.age = NULL, strata = NULL, probabilities = NULL, rates = NULL,
  root.edge = TRUE, use.exact.times = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{species}{Taxonomy object.}

\item{interval.ages}{Vector of stratigraphic interval ages, starting with the minimum age of the youngest interval and ending with the maximum age of the oldest interval.}

\item{basin.age}{Maximum age of the oldest stratigraphic interval.}

\item{strata}{Number of stratigraphic intervals.}

\item{probabilities}{Probability of sampling/preservation in each interval. The number of probabilities should match the number of intervals.}

\item{rates}{Poisson sampling rate for each interval. The number of rates should match the number of intervals.}

\item{root.edge}{If TRUE include the root edge. Default = TRUE.}

\item{use.exact.times}{If TRUE use exact sampling times. If FALSE hmin and hmax will equal the start and end times of the corresponding interval. Default = TRUE.}
}
\value{
An object of class fossils.
}
\description{
Intervals can be specified by specifying the interval boundaries using \code{interval.ages} or specifying both \code{basin.age} and \code{strata}.
In the second scenario all intervals will be of equal length.
Preservation can be specified using \code{rates}, which represent the rates of a Poisson process in each interval,
or \code{probabilities}, which represent the probabilities of sampling per interval.
When using \code{probabilities}, at most one fossil per species will be sampled per interval.
}
\examples{

# simulate tree
t <- ape::rtree(6)

# assign a max age based on tree height
max.age <- basin.age(t)

# simulate fossils using basin.age and strata & probabilities
strata = 4
probability = rep(0.7, 4)
f <- sim.fossils.intervals(t, basin.age = max.age, strata = strata, probabilities = probability)
plot(f, t, strata = strata, show.strata = TRUE)

# simulate fossils using interval.ages & rates
times = seq(0, max.age, length.out = 4)
rates = c(5, 0, 5)
f <- sim.fossils.intervals(t, interval.ages = times, rates = rates)
plot(f, t)

# simulate fossils using taxonomy
s <- create.taxonomy(t, 0.1, 0.1, 1)
f <- sim.fossils.intervals(species = s, interval.ages = times, rates = rates)
plot(f, t)

}
\keyword{fossil}
\keyword{non-uniform}
\keyword{preservation}
\keyword{uniform}
