% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.non.unif.depth}
\alias{sim.fossils.non.unif.depth}
\title{Simulate fossils under a non-uniform model of preservation (Holland, 1995)}
\usage{
sim.fossils.non.unif.depth(tree, profile, PA = 0.5, PD = 0.5, DT = 0.5,
  interval.ages = NULL, basin.age = NULL, strata = NULL,
  root.edge = TRUE, convert.rate = FALSE)
}
\arguments{
\item{tree}{Phylo object.}

\item{profile}{Vector of relative water depth. The first number corresponds to the youngest interval. The length of the vector should 1 less than the length of interval.ages.}

\item{PA}{Peak adbundance parameter.}

\item{PD}{Preferred depth parameter.}

\item{DT}{Depth tolerance parameter.}

\item{interval.ages}{Vector of stratigraphic interval ages, starting with the minimum age of the youngest interval and ending with the maximum age of the oldest interval.}

\item{basin.age}{Maximum age of the oldest stratigraphic interval.}

\item{strata}{Number of stratigraphic intervals.}

\item{root.edge}{If TRUE include the root edge (default = TRUE).}

\item{convert.rate}{If TRUE convert per interval sampling probability into a per interval Poisson rate (default = FALSE).}
}
\value{
An object of class fossils.
sp = edge labels. h = fossil or interval ages. If convert.rate = TRUE, h = specimen age, if convert.rate = FALSE, h = max horizon age.
}
\description{
This function uses a three parameter Guassion model to simulate non-uniform fossil preservation along a specified phylogeny.
Preservation varies with respect to water depth, which is used as a proxy for changes in sedimentary environment.
The per interval probability of sampling is \deqn{P(collection) = PA e ^ (-(d - PD)^2 / 2*DT^2 ) }
where \emph{PA} is species peak abundance, \emph{PD} is preferred depth, \emph{DT} is depth tolerance and \emph{d} is current water depth.
\emph{PD} is the depth at which the species is most likely to be found and is equivalent to the mean of the distribution.
\emph{PA} is the probability of sampling an occurrence at this depth.
\emph{DT} is the potential of a species to be found at a range of depths and is equivalent to the standard deviation. \cr \cr
Non-uniform interval ages can be specified as a vector (\code{interval.ages}) or a uniform set of interval ages can be specified using
maximum interval age (\code{basin.age}) and the number of intervals (\code{strata}), where interval length \eqn{= basin.age/strata}.
}
\examples{
# simulate tree
t<-ape::rtree(6)
# assign a max age based on tree height
max = basin.age(t)
# generate water depth profile
strata = 7
wd<-sim.water.depth(strata)
# simulate fossils
f<-sim.fossils.non.unif.depth(t, wd, PA = 1, PD = 0.5, DT = 1, basin.age = max, strata = strata, convert.rate = TRUE)
plot(f,t, show.proxy = T, proxy.data = wd, strata = strata, show.strata = T)
}
\references{
Holland, S.M. 1995. The stratigraphic distribution of fossils. Paleobiology 21: 92–109.
}
\keyword{fossil}
\keyword{non-uniform}
\keyword{preseravtion}

