% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.non.unif.depth}
\alias{sim.fossils.non.unif.depth}
\title{Simulate fossils under a non-uniform model of preservation (Holland, 1995)}
\usage{
sim.fossils.non.unif.depth(tree = NULL, species = NULL,
  interval.ages = NULL, basin.age = NULL, strata = NULL,
  depth.profile = NULL, PA = 0.5, PD = 0.5, DT = 0.5,
  use.rates = FALSE, root.edge = TRUE, use.exact.times = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{species}{Taxonomy object.}

\item{interval.ages}{Vector of stratigraphic interval ages, starting with the minimum age of the youngest interval and ending with the maximum age of the oldest interval.}

\item{basin.age}{Maximum age of the oldest stratigraphic interval.}

\item{strata}{Number of stratigraphic intervals.}

\item{depth.profile}{Vector of relative water depth. The first number corresponds to the youngest interval. The length of the vector should be 1 less than the length of interval.ages.}

\item{PA}{Peak abundance parameter.}

\item{PD}{Preferred depth parameter.}

\item{DT}{Depth tolerance parameter.}

\item{use.rates}{If TRUE convert per interval sampling probability into a per interval Poisson rate. Default = FALSE.}

\item{root.edge}{If TRUE include the root edge. Default = TRUE.}

\item{use.exact.times}{If TRUE use exact sampling times. If FALSE \code{hmin} and \code{hmax} will equal the start and end times of the corresponding interval. Default = TRUE.}
}
\value{
An object of class fossils.
}
\description{
This function uses a three parameter Gaussian model to simulate non-uniform fossil preservation along a specified phylogeny.
Preservation varies with respect to water depth, which is used as a proxy for changes in sedimentary environment.
The per interval probability of sampling is \deqn{P(collection) = PA e ^ (-(d - PD)^2 / 2*DT^2 ) }
where \emph{PA} is species peak abundance, \emph{PD} is preferred depth, \emph{DT} is depth tolerance and \emph{d} is current water depth.
\emph{PD} is the depth at which the species is most likely to be found and is equivalent to the mean of the distribution.
\emph{PA} is the probability of sampling an occurrence at this depth.
\emph{DT} is the potential of a species to be found at a range of depths and is equivalent to the standard deviation. \cr \cr
Non-uniform interval ages can be specified as a vector (\code{interval.ages}) or a uniform set of interval ages can be specified using
maximum interval age (\code{basin.age}) and the number of intervals (\code{strata}), where interval length \eqn{= basin.age/strata}. \cr \cr
Simulate fossils for a phylo (\code{tree}) or taxonomy object (\code{species}).
If both are specified, the function uses taxonomy.
If no taxonomic information is provided, the function assumes all speciation is symmetric (i.e. budding, \code{beta = 1}).
}
\examples{
# simulate tree
t = ape::rtree(6)

# assign a max age based on tree height
max.age = basin.age(t)

# generate water depth profile
strata = 7
wd = sim.water.depth(strata)

# simulate fossils using tree & basin.age and strata
f = sim.fossils.non.unif.depth(t, basin.age = max.age, strata = strata,
 depth.profile = wd, PA = 1, PD = 0.5, DT = 1, use.rates = TRUE)
plot(f,t, show.proxy = TRUE, proxy.data = wd, strata = strata, show.strata = TRUE)

# simulate fossils using taxonomy & interval.ages
s = create.taxonomy(t, 0.1, 0.1, 1)
times = seq(0, max.age, length.out = strata + 1)
f = sim.fossils.non.unif.depth(species = s, interval.ages = times,
     depth.profile = wd, PA = 1, PD = 0.5, DT = 1, use.rates = TRUE)
plot(f,t)

}
\references{
Holland, S.M. 1995. The stratigraphic distribution of fossils. Paleobiology 21: 92-109.
}
\seealso{
\code{\link{sim.fossils.poisson}}, \code{\link{sim.fossils.intervals}}
}
\keyword{fossil}
\keyword{non-uniform}
\keyword{preseravtion}
