% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.non.unif}
\alias{sim.fossils.non.unif}
\title{Simulate fossils under a non-uniform model of preservation (Holland, 1995)}
\usage{
sim.fossils.non.unif(tree, basin.age, strata, profile, PA = 0.5, PD = 0.5,
  DT = 0.5, root.edge = T, convert.rate = F)
}
\arguments{
\item{tree}{Phylo object.}

\item{basin.age}{Maximum age of the oldest stratigraphic interval.}

\item{strata}{Number of stratigraphic horizons.}

\item{PA}{Peak adbundance parameter.}

\item{PD}{Preferred depth parameter.}

\item{DT}{Depth tolerance parameter.}

\item{root.edge}{If TRUE include the root edge (default = FALSE).}

\item{convert.rate}{If TRUE convert per interval sampling probability into a per interval Poisson rate (default = FALSE).}
}
\value{
dataframe of sampled fossils.
sp = edge labels. h = fossil or horizon ages. If convert.rate = TRUE, h = specimen age, if convert.rate = FALSE, h = max horizon age.
}
\description{
Simulate fossils under a non-uniform model of preservation (Holland, 1995)
}
\keyword{fossil}
\keyword{preseravtion}
\keyword{uniform}

