% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.non.unif}
\alias{sim.fossils.non.unif}
\title{Simulate fossils under a non-uniform model of preservation (Holland, 1995)}
\usage{
sim.fossils.non.unif(tree, basin.age, strata, profile, PA = 0.5, PD = 0.5,
  DT = 0.5, root.edge = T)
}
\arguments{
\item{tree}{Phylo object.}

\item{basin.age}{Maximum age fossils may be sampled.}

\item{strata}{Number of stratigraphic horizons.}

\item{PA}{Peak adbundance parameter.}

\item{PD}{Preferred depth parameter.}

\item{DT}{Depth tolerance parameter.}

\item{root.edge}{Boolean indicating tree inclues a root edge.}
}
\value{
dataframe of sampled fossils.
sp = edge labels. h = horizon labels (= max horizon age).
}
\description{
Simulate fossils under a non-uniform model of preservation (Holland, 1995)
}
\keyword{fossil}
\keyword{preseravtion}
\keyword{uniform}

