% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.poisson}
\alias{sim.fossils.poisson}
\title{Simulate fossils under a Poisson sampling model}
\usage{
sim.fossils.poisson(tree, rate, root.edge = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{rate}{Poisson sampling rate.}

\item{root.edge}{If TRUE include the root edge (default = TRUE).}
}
\value{
An object of class fossils.
sp = edge labels. h = ages.
}
\description{
Simulate fossils under a Poisson sampling model
}
\examples{
# simulate tree
t<-ape::rtree(4)
# simulate fossils
rate = 2
f<-sim.fossils.poisson(t, rate)
plot(f, t)
}
\keyword{preseravtion}
\keyword{uniform}

