% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.poisson}
\alias{sim.fossils.poisson}
\title{Simulate fossils under a Poisson sampling model}
\usage{
sim.fossils.poisson(rate, tree = NULL, species = NULL, root.edge = TRUE)
}
\arguments{
\item{rate}{Poisson sampling rate.}

\item{tree}{Phylo object.}

\item{species}{Taxonomy object.}

\item{root.edge}{If TRUE include the root edge. Default = TRUE.}
}
\value{
An object of class fossils.
}
\description{
Simulate fossils for a phylo (\code{tree}) or taxonomy object (\code{species}).
If both are specified, the function uses taxonomy.
If no taxonomic information is provided, the function assumes all speciation is symmetric (i.e. \code{beta = 1}).
}
\examples{

# simulate tree
t <- ape::rtree(6)

# simulate fossils using the tree
rate = 2
f <- sim.fossils.poisson(rate, tree = t)
plot(f, t)

# simulate fossils using taxonomy
s <- create.taxonomy(t, 0.5, 1, 0.5)
f <- sim.fossils.poisson(rate, species = s)
plot(f, t)

}
\keyword{preservation}
\keyword{uniform}
