% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.som.R
\name{sim.fossils.poisson.rh}
\alias{sim.fossils.poisson.rh}
\title{Simulate fossils under a Poisson sampling model with individual rates for every edge}
\usage{
sim.fossils.poisson.rh(tree, rd = rlnorm(nrow(tree$edge) + 1, 0.1, 0.1),
  root.edge = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{rd}{The desired rate distribution from which the rates for individual edges would be
sampled (default = rlnorm(nrow(t$edge)+1,0.1,0.1)).
The total number of rates to be sampled is equal to the number of edges (include the root
edge) in the tree.}

\item{root.edge}{If TRUE include the root edge (default = TRUE).}
}
\value{
A list with an object of class fossils and a vector of rates applied on each edge.
sp = node labels. h = ages. rates = rates.
}
\description{
Simulate fossils under a Poisson sampling model with individual rates for every edge
}
\examples{
# simulate tree
t<-TreeSim::sim.bd.taxa(4,1,1,0.10)[[1]]
# simulate fossils
f<-sim.fossils.poisson.rh(t)
plot(f$fossils, t)
f$rates
}
