% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.unif}
\alias{sim.fossils.unif}
\title{Simulate fossils under a uniform model of preservation}
\usage{
sim.fossils.unif(tree, basin.age, strata, sampling, root.edge = T,
  convert.rate = FALSE)
}
\arguments{
\item{tree}{Phylo object.}

\item{basin.age}{Maximum age of the oldest stratigraphic interval.}

\item{strata}{Number of stratigraphic horizons.}

\item{sampling}{Probability of sampling/preservation.}

\item{root.edge}{If TRUE include the root edge (default = FALSE).}

\item{convert.rate}{If TRUE convert per interval sampling probability into a per interval Poisson rate (default = FALSE).}
}
\value{
dataframe of sampled fossils.
sp = edge labels. h = fossil or horizon ages. If convert.rate = TRUE, h = specimen age, if convert.rate = FALSE, h = max horizon age.
}
\description{
Simulate fossils under a uniform model of preservation
}
\examples{
t<-ape::rtree(4)
ba<-basin.age(t,root.edge=F)
sim.fossils.unif(t,ba,5,0.5)
}
\keyword{fossil}
\keyword{preseravtion}
\keyword{uniform}

