% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.unif}
\alias{sim.fossils.unif}
\title{Simulate fossils under a uniform model of preservation}
\usage{
sim.fossils.unif(tree, basin.age, strata, sampling)
}
\arguments{
\item{tree}{Phylo object.}

\item{basin.age}{Maximum age fossils may be sampled.}

\item{strata}{Number of stratigraphic horizons.}

\item{sampling}{Probability of sampling/preservation.}
}
\value{
dataframe of sampled fossils.
sp = edge labels. h = horizon labels (= max horizon age).
}
\description{
Simulate fossils under a uniform model of preservation
}
\examples{
t<-rtree(4)
sim.fossils.unif(t,1,5,0.5)
}
\keyword{preseravtion}
\keyword{uniform}

