% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.unif}
\alias{sim.fossils.unif}
\title{Simulate fossils under a uniform model of preservation for a set of equal length intervals}
\usage{
sim.fossils.unif(tree, basin.age, strata, probability, root.edge = T,
  convert.rate = FALSE, use.exact.times = TRUE)
}
\arguments{
\item{tree}{Phylo object.}

\item{basin.age}{Maximum age of the oldest stratigraphic interval.}

\item{strata}{Number of stratigraphic intervals.}

\item{probability}{Probability of sampling/preservation.}

\item{root.edge}{If TRUE include the root edge (default = TRUE).}

\item{convert.rate}{If TRUE convert per interval sampling probability into a per interval Poisson rate (default = FALSE).}
}
\value{
An object of class fossils.
}
\description{
Simulate fossils under a uniform model of preservation for a set of equal length intervals
}
\examples{
# simulate tree
t<-ape::rtree(6)
# assign a max age based on tree height
max<-basin.age(t)
# simulate fossils
strata = 4
probability = 0.7
f<-sim.fossils.unif(t, max, strata, probability)
plot(f, t, binned = TRUE, strata = strata)
}
\keyword{fossil}
\keyword{preservation}
\keyword{uniform}
