% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fossils.R
\name{sim.fossils.unif}
\alias{sim.fossils.unif}
\title{Simulate fossils under a uniform model of preservation}
\usage{
sim.fossils.unif(tree, basin.age, strata, sampling, root.edge = T)
}
\arguments{
\item{tree}{Phylo object.}

\item{basin.age}{Maximum age fossils may be sampled.}

\item{strata}{Number of stratigraphic horizons.}

\item{sampling}{Probability of sampling/preservation.}
}
\value{
dataframe of sampled fossils.
sp = edge labels. h = horizon labels (= max horizon age).
}
\description{
Simulate fossils under a uniform model of preservation
}
\examples{
t<-ape::rtree(4)
ba<-basin.age(t,root.edge=F)
sim.fossils.unif(t,ba,5,0.5)
}
\keyword{fossil}
\keyword{preseravtion}
\keyword{uniform}

