% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.trait.values.R
\name{sim.water.depth}
\alias{sim.water.depth}
\title{Simulate water depth profile}
\usage{
sim.water.depth(strata, depth = 2, cycles = 2)
}
\arguments{
\item{strata}{Number of stratigraphic intervals}

\item{depth}{Maximum water depth.}

\item{cycles}{Number of cycles (transgressions and regressions).}
}
\value{
dataframe of sampled water depths.
}
\description{
Function returns water depth profile using the sine wave function \eqn{y = depth*sin(cycles*pi*(x-1/4))}.
}
\examples{
strata = 100
wd = sim.water.depth(strata)
plot(wd, type="l")
}
\keyword{fossil}
\keyword{non-uniform}
\keyword{preservation}
